fix(queue): finalize items 1-2 reliability and observability wrap-up
- task_queue: receive() with timeout_seconds<=0 now polls without blocking
  (asyncio.wait_for(...,0) previously always timed out); widen timeout types
  to float so short test timeouts work while Redis blpop keeps int/float parity
- task_worker: accept float block_timeout_seconds
- tests: align queue metrics assertions with distinct in-flight task semantics
  (Redis hlen parity); fix await/subscript precedence in ack test; refresh
  trace fixture timestamp so TraceStore TTL pruning keeps it; health test now
  expects the task_queue key added by queue monitoring
- docker-compose: fix dead-letter env indentation under api and worker
- launcher: 打开/关闭芯智导学.cmd delegate to scripts/xzd_supervisor.ps1
  (auto-restart supervisor + manual stop entry)

+++ apps/api/app/services/task_worker.py
@@ -23,7 +23,7 @@ class TaskWorker:
         dispatcher: TaskDispatcher,
         queue: TaskQueue,
         *,
-        block_timeout_seconds: int,
+        block_timeout_seconds: float,
         recovery_interval_seconds: int,
     ) -> None:
         self.dispatcher = dispatcher
